requirements.txt
requests==2.23.0
beautifulsoup4==4.8.2
Django==3.0.3
djangorestframework==3.11.0
pandas==1.0.1
